app/BackEnd/requirements.txt
Pillow==8.2.0
boto3==1.17.69
pandas==1.1.5
numpy==1.19.5
PyYAML==5.4.1
fastapi[all]
uvicorn

